Moved markdown parser to base
